Refactor new module animatronic with pygame and serial
