+++ code/python/animatronic/serialHat.py
@@ -11,11 +11,11 @@ def open_serial_port(com_port, baudrate):
         time.sleep(2)  # Esperar a que Arduino se reinicie y establezca la conexión
         if ser.in_waiting:
             ser.read(ser.in_waiting) # Vaciar buffer de entrada
-        logging.info(f"Puerto serie '{com_port}' abierto correctamente.")
+        logger.info(f"Puerto serie '{com_port}' abierto correctamente.")
         print(f"Puerto serie '{com_port}' abierto correctamente.")
         return ser
     except serial.SerialException as e:
-        logging.error(f"Error al abrir el puerto '{com_port}': {e}")
+        logger.error(f"Error al abrir el puerto '{com_port}': {e}")
         print(f"Error: No se pudo abrir el puerto '{com_port}'. Asegúrate de que Arduino esté conectado y el puerto sea correcto.")
         print(f"Detalle del error: {e}")
         return None
@@ -26,7 +26,7 @@ def send_servo_command(ser_port, command):
         try:
             message = f"{command}\n" # Añade salto de línea para que Arduino sepa cuándo termina el comando
             ser_port.write(message.encode()) # Codifica el mensaje a bytes
-            logging.debug(f"Enviado a Arduino: {message.strip()}") # strip() para no loguear el '\n'
+            logger.debug(f"Enviado a Arduino: {message.strip()}") # strip() para no loguear el '\n'
             if command == '1':
                 print("Servo: MOVING")
             elif command == '0':
@@ -34,18 +34,18 @@ def send_servo_command(ser_port, command):
             time.sleep(0.02) # Pequeño delay después de enviar para dar tiempo a Arduino a procesar
             return True
         except Exception as e:
-            logging.error(f"Error al enviar comando '{command}' a Arduino: {e}", exc_info=True)
+            logger.error(f"Error al enviar comando '{command}' a Arduino: {e}", exc_info=True)
             print(f"Error crítico de comunicación con Arduino: {e}")
             return False
     else:
-        logging.warning(f"Intento de enviar comando '{command}', pero el puerto serial no está abierto.")
+        logger.warning(f"Intento de enviar comando '{command}', pero el puerto serial no está abierto.")
         return False
 
 def close_serial_port(ser_port):
     """Cierra el puerto serial si está abierto."""
     if ser_port and ser_port.is_open:
         ser_port.close()
-        logging.info("Puerto serie cerrado.")
+        logger.info("Puerto serie cerrado.")
         print("Puerto serie cerrado.")
 
 def read_arduino_feedback(ser_port):
@@ -77,7 +77,7 @@ def read_arduino_feedback(ser_port):
                         print(f"    [Servo Info]: S1 a: {angle1}°, S2 a: {angle2}° (Movimiento aleatorio)")
                 # Add more conditions for other Arduino debug messages if needed
                 
-                logging.debug(f"Recibido de Arduino: {feedback}")
+                logger.debug(f"Recibido de Arduino: {feedback}")
         except Exception as e:
-            logging.error(f"Error al leer feedback de Arduino: {e}", exc_info=True)
+            logger.error(f"Error al leer feedback de Arduino: {e}", exc_info=True)
             print(f"[ARDUINO ERROR]: {e}")
